Bound the response body handed to OnResponseReceived in idunno.AtProto (#453)
Requests are made with HttpCompletionOption.ResponseHeadersRead, so an
OnResponseReceived handler was given an unbuffered network stream. A handler
which read it allocated whatever the service chose to send, bypassing
MaximumResponseSize entirely. Verified against a test server: with a 32MB cap
and a 48MB response a handler read the full 48MB before the library reported
ResponseTooLarge.

The body is now buffered under MaximumResponseSize before the handler is
called, and an over-large response is rejected without the handler seeing it.
Buffering only happens when a handler has been attached, so the streaming path
is unchanged when nothing is going to read the body.

Also:

* AtProtoHttpClient.OnSendingRequest and OnResponseReceived are now settable.
  Both were documented as "Gets or sets" but only had a getter, so no handler
  could be attached through the non-generic client at all. Setting either to
  null now throws rather than leaving a client which throws on send.
* The handlers are held on the non-generic client and applied when its inner
  client is constructed. Setting them through the inner client would have
  created it before an object initializer had the chance to run the
  MaximumResponseSize init accessor, silently keeping the default limit.
* HttpContentReader grows its read buffer through long arithmetic. Doubling in
  int overflowed to a negative length when MaximumResponseSize was above 1GB,
  throwing from ArrayPool.Rent instead of reporting ResponseTooLarge.

+++ src/idunno.AtProto/HttpContentReader.cs
@@ -110,7 +110,9 @@ internal static class HttpContentReader
                 {
                     if (bytesRead == buffer.Length)
                     {
-                        byte[] grown = ArrayPool<byte>.Shared.Rent(Math.Min(buffer.Length * 2, maximumLength));
+                        // Grow through a long, as doubling a buffer larger than half of int.MaxValue would otherwise
+                        // overflow to a negative length and throw from Rent rather than stopping at maximumLength.
+                        byte[] grown = ArrayPool<byte>.Shared.Rent((int)Math.Min((long)buffer.Length * 2, maximumLength));
                         Buffer.BlockCopy(buffer, 0, grown, 0, bytesRead);
                         ArrayPool<byte>.Shared.Return(buffer);
                         buffer = grown;
